fill_all_forms updated and new log data
Updated the fill_all_forms function to be based on a list of functions put in order and added some index (WIP) to continue when problems are encountered. In addition learning a little more about logging and so changed some of logging calls to provide better info. Still buggy, if element not in view will fail.